requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
python-dotenv==1.0.0
httpx==0.24.1
pydantic==1.10.13
python-json-logger==2.0.7 